converted log to variable entries and enable striping
changed the unit-server to support variable entry sizes. enabled configuration with multiple stripes and multiple servers
(this was just a matter of allowing configurations to specify more than one group, everything else was supported, essentially)

this means that entries are appended to the storage unit in the order
they arrive, not the logical order of the log. This also means that
trimming a prefix of the log may result in trimming less entries than
requested, because we only trim contiguous space