Set router's epoch at LayoutServer's reboot()
I've looked for a couple of hours this morning trying to figure out
how the slf/quickcheck-testing1a branch (before Replex merge to
master, before I went on vacation) is able to change the NettyServerRouter's
epoch to 0 after reset() & reboot(), but this branch does not.
I'm tired of this fiddly & brittle code, so I'm putting something big
and obvious into reboot() to make it clear where & when setServerEpoch()
ought to be called.