# This is a combination of 4 commits. # The first commit's message is: Fixed a concurrency issue in LogUnitServer. For some reason, Caffeine may invoke handleEviction multiple times on the same entry. Added check that refCnt() > 0 before calling release().
# The 2nd commit message will be skipped:

#	Further trying to distill a test that determinitically surfaces the concurrency
#	issue.

# The 3rd commit message will be skipped:

#	This commit fixes a buffer management race issue, which has been introduced into
#	Corfu during the introduction of the StreamLog abstraction.
#
#	The probem was as follows:
#
#	In LogUnitServer, when the Caffeine cache evicts an entry (see
#	LogUnitServer::handleEviction), LogUnitServer called
#	release() on this entry. However, when in-memory mode, the entry is still used
#	in the InMemoryStreamLog. By releasing it, we risk re-allocating the memory and overwriting it.
#
#	Furthermore, when Caffeine re-retrieved the entry into the cache (see
#	LogUnitServer::handleRetrieval), the in-memory log entry was found and returned,
#	but later, Caffine again evicted it, re-releasing an entry that had refCnt zero!
#
#	The fix is to let StreamLlog handle release requests explicitly: an in-memory
#	StreamLog does nothing, while a normal StreamLog actually releases the buffer.
#
#	merge

# This is the 4th commit message:

This commit fixes a buffer management race issue, which has been introduced into
Corfu during the introduction of the StreamLog abstraction.

The probem was as follows:

In LogUnitServer, when the Caffeine cache evicts an entry (see
LogUnitServer::handleEviction), LogUnitServer called
release() on this entry. However, when in-memory mode, the entry is still used
in the InMemoryStreamLog. By releasing it, we risk re-allocating the memory and overwriting it.

Furthermore, when Caffeine re-retrieved the entry into the cache (see
LogUnitServer::handleRetrieval), the in-memory log entry was found and returned,
but later, Caffine again evicted it, re-releasing an entry that had refCnt zero!

The fix is to let StreamLlog handle release requests explicitly: an in-memory
StreamLog does nothing, while a normal StreamLog actually releases the buffer.