This commit fixes the buffer management race issue, which I think has been introduced into Corfu during the introduction of the StreamLog abstraction.
The probem was as follows:

In LogUnitServer, when the Caffeine cache evicts an entry (see
LogUnitServer::handleEviction), LogUnitServer called
release() on this entry. However, when in-memory mode, the entry is still used
in the InMemoryStreamLog. By releasing it, we risk re-allocating the memory and
overwriting it.

Furthermore, when Caffeine re-retrieved the entry into the cache (see
LogUnitServer::handleRetrieval), the in-memory log entry was found and returned,
but later, Caffine again evicted it, re-releasing an entry that had refCnt zero!

The fix is to let StreamLog, which is the last component that handles the entry,
handle release requests explicitly. The in-memory StreamLog does nothing on
release, a normal StreamLog actually releases the buffer.