Opacity (#390)
* 1. further touches to interleaving scheduler.
2. the test concurrentAbortMultiViewTest did not really look at the same map;
probably not the intention. Modified it, and also produced an
interleaved-schedule variant.

* still dealing with PARAMETERS issues

* state machine tests can now run either with controled interleaving of
state-machine steps, or in a multi-threaded mode.

In a multi-threaded mode, each state-machine instance is run to completion
inside a thread. The concurrency level is determined, as usual, by setting the
number of threads.

In this manner, the same test can now run with thread-controled interleaving,
or under the interleaving-scheduler control.

* rebase leftovers

* added utils for building test state machines

* fixed tests

* running concurrency test in threaded mode revealed a race in the test (which
does not occur in interleaved mode, since it needs to switch in the middle of a
state-machine "step"). Fixed in OptimisticTX.

TODO: need to verify the same race in WriteWriteTX.

* fixed WriteWriteTX test.