More fixes to the gobuild target
After the module change the .deb files were being generated in
a different  location, updated the gobuild target to publish the new
.deb.

Also, for some reason before we run mvn deploy we run a sed command to
add the build number to the pom file. Because maven depdendency
resolution depends on the file name and child modules depened on
on the root pom the build is breaking.

Change-Id: I37a436dcdb7f227f235928f3777a418a66aae689