stream-based conflict parameter tracking (#414)
* worked out write-write conflict resolution and started to build tests of
performance and abort rates.

* organized some of the TX concurrency tests. still WIP.

* more test organization

* added a samples package with tutorials. Will keep adding tutorials; inviting
everyone to look at existing ones and comment on them, as well as add their own!

* addressing style issues in samples.

* letter swap

* samples

* repeatable stream test with a potential problem

* addressing stylistic changes requested by reviewer

* stylistic changes to sample program

* snapshot TX

* cleanup tests

* demonstrating hole-race

* added samples and various cleanups

* removed printouts, and a few minor startup cleanups for tests.

* minor stylistic amendments.

* added a test that demonstrating a current bug in the implementation of
VersionLockedObject.

The required fix (todo):

When a transasction commits, it needs to sync all the streams in its write-set up to the transaction
point.

* demonstrating concurrency via a two-runtime test.

* getting travis to pass.

* patching up concurrency tests

* minor cleanups, removing printouts mainly

* refined the confict parameter tracking to include information on conflict
parameters -per-stream.

This should result in slightly improved performance if the same conflict
parameter value(s) are used across streams.

For example, say that the same key "K" is put into two maps, map1, and map2, by
simultaneous transactions. There should be no conflict. In constrast, if "K" is
inserted twice to the same map within concurrent transactions, they should
conflict (in write-write conflict mode).

By tracking not only the parameter value but also the stream name, we avoid
false conflict in the first case.

* removed sequencer lease.

* Made conflictKey its own class instead of a SimpleEntry pair.