First draft for the quorum replication
* Writer and fillHole send write request to all servers, but wait only n/2+1 servers respond.
* Reader queries all the servers, but continues after the first successful response.
* The logic how to use the futures is encapsulated in a new QuorumReplicationFuture. This future does not implements any locking itself, it  just delegates the locking to the leafs.  Also no any threads are added manually, just the internally CompletableFuture leafs use the ManagedBlocker in their ForkJoinPool.  I believe this approach is novel, at least I don't see anything similar in the public code repos. This helps us not only with the testability, but also we also keep the replication view readable.

All tests are not ignored & green, including the excellent test StreamViewTest.streamCanSurviveOverwriteException().

Non-steady states: In case of the server failures, this approach by concept will be more solid, as it will continue to read and write successfully (the chain implementation would block during the write).
Even if everything looks working now, this implementation is a "draft" as the behaviour when the sequencer fails needs more attention (+ dedicated tests).  This is exactly what I am doing now.