Getting other runtimes to the correct epoch/layout after a layout change
When a layout change is triggered by a client, this client will drive
paxos to run on the servers to establish a new layout in a new epoch. It will
also update itself to the new layout/epoch since he is the driver of it.

Other clients (other runtimes) will not be signaled of the changes. The next time
they will ask for a token or try to read/write, they will receive a WrongEpoch
exception.

This fix will implement the logic that bring the runtime to the new epoch/layout.
Before the fix, other runtimes where stuck in a infinite loop of receiving WrongEpoch
exception.