Read quorum - only if quorum from the same values is reached, changes in the future from the review request
This change is just about the address the reviews from  Dalia, Kevin and Medhavi, NOT about recovery-write/2 step hole fill, that hopefully will come next (later tonight?).

The change corrects the future and allows conflicting results in it. The reason - when using this future to read (https://github.com/CorfuDB/CorfuDB/wiki/Quorum-Replication) we need a quorum of the SAME values, not just quorum of any value. The reason is that during the step 2 of the recovery write, there is possibility of conflicting values (if a quorum is reached at step 1, then a client with a higher rank reaches a quorum, the first client might still be able to write to some of the servers).

Now in the future we counting the values, for lighter implementation we
are reusing the TreeMultimap from Guava. We are comparing by value the
Read quorum - only if quorum from the same values is reached, changes in
the future from the review request

This change is just about the address the reviews from  Dalia, Kevin and
Medhavi, NOT about recovery-write/2 step hole fill, that hopefully will
come next (later tonight?).

The change corrects the future and allows conflicting results in it. The
reason - when using this future to read
(https://github.com/CorfuDB/CorfuDB/wiki/Quorum-Replication) we need a
quorum of the SAME values, not just quorum of any value. The reason is
that during the step 2 of the recovery write, there is possibility of
conflicting values (if a quorum is reached at step 1, then a client with a
higher rank reaches a quorum, the first client might still be able to
write to some of the servers).

Now in the future we counting the values, for lighter implementation we
are reusing the TreeMultimap from Guava. We are comparing by value the
ReadResponses by serializing them, which is not so quick, but as it is
used only rarery, I don't believe it will be a problem. If there is no way to reach a decision, an exceptio is thrown.