Quorum replication, code/feature complete milestone
Sending you the code just for review, as I want to show how I implement all aspects of the quorum replication.  If we are OK with my approach, the next submission will contain
better unit test coverage (still working on it),  minor bugfixes, extra logging, reactions to review comments, and... that's it.

So - this implementation was a real adventure.

IN ODRER TO IMPLEMENT THE SPEC OF DALIA, I DID:

https://github.com/CorfuDB/CorfuDB/wiki/Quorum-Replication

1) A separate stream view (in the style of Michael) that is able to change position, and it keeps and increases the rank.

2) The replication view is able to send different hole and write messages for the different phases.

3) So we have new WriteModes enums for the WriteRequest message.

4) In the same way we have a new new FillHole mode enum in FillHlleRequest message. This is the only change, that concerns the existing chain replication mode, but a 2 new bytes for this rare message should not be a problem.

5) Two new data types are defined DATA_PROPOSED and HOLE_PROPOSED, but these 2 data types never go to the log, they are only in the in-memory cache in quorum delegate
for the log server (see below)

6) On the server side: the replex and new quorum are extracted to delegates in order not to bloat the default implementation

7) The quorum implementation keeps the DATA_PROPOSED and HOLE_PROPOSED messages in internal cache. I permit this cache to expire to prevent DoS.

8) The business logic with the comparing of the ranks from the spec is in the quorum delegate

9) A synchronization/locking is needed here, as our updates are needed on reads. But my implementation is smart enough not to look every write message, but only the message on the specified address (see PrefixBasedLocalLocks class, I believe it's  quite interesting approach. ).

REMARK: The non steady state is detected by the client in two ways
1) Write/read conflicts between the values.
2) By new method in the sequencer, which is currently just for unit testing. Making the sequencer changeable and having a real backup sequencer  as specified by Dalia is is different efort.

END OF THE EXPECTED FEATURES. LET'S HAVE A STEP BACK:

WHAT WAS NOT IN THE DOCUMENT OF DALIA, BUT I NEEDED TO IMPLEMENT:
 - Step3 (optional,  asynchronous) - overriding of conflicting values  after successful write or hole fill. Luckily there are no such values, but doing this was the most difficult task.

ONE MORE STEP BACK - I realized a few days ago that I must do this in order to have a production level quality.  The issue something we already discussed, when we were working on the "Future class -
quorum algorithm gives warranty that after step 2 we will have a quorum of the same values, but  it DOES NOT give warranty, that the some different values won't be stored. So - imagine 3 servers, quorum of 2 with value A and sinle server with value B. One of the first 2 server dies, and we will have only A against B - the quorum is lost.

In order to attack this, after the write gains quorum,  a bit later after the write, the scheduler quickly checks the status of the futures. If therese is still something unresolved outside of the quorum, an override request is scheduled.

Sounds simple and quick. But... on the server side I was badly surprised - I needed to deal with the override exception (and in general with the fact the override concept was not natural).  So what I am doing - I am not changing log entities (we continue to work in append only manner), I am just trimming the previous entities by maintaining a file of blacklisted indexes.

I know the last can overlap with the efforts to implement trim, but I believe I will be able to merge this into different implementation too, if necessary.  The GC can clean these "trim index" files.