change JRE and debian package name for corfudb
Previously corfudb use java-runtime-headless as its default
runtime. It is changed to oracle-java8-jre to comply with nsx-mp.
Since nsx-mp consumes corfudb as a debian package, it makes more
sense to be named as "nsx-corfu" rather than the default name
"infrastructure".
The version of nsx-corfu is set to 0.1.0 regarding their original
"0.1-SNAPSHOT".

Issue: #
Change-Id: Ib9421c2d632414fa355859e986b60fdd87c06d11