Fix the version after undo operation in rollbackUnsafe
Previously, once we were undoing a record, we would leave the object at the version of the
record we undid.
This would trigger a resetObjectUnsafe, because once we undo the latest record, we will check that
the current version of the object is <= version we wanted to unroll to. This would not be the case
and would iterate again through the while loop, generating a NoRollBackException (undoLog is empty).

In this commit, we use the backpointermap to get the previous version of the object and update the
version to its predecessor.