trivial change to force nsx-corfudb rebuild
The current recommended build of nsx-corfudb is too old.
It will make all mp-review-test fail with recommended build
after nsx-corfudb is added to appliance.

This dummy change does nothing more than trigger a rebuild in
gobuild system. It helps to update the recommended build of
corfu in nsx-transformer after passing the ToT colibri test.

This is a temporary action. In the future we may need to able
override the recommended build manually.

Issue: #
Change-Id: I0c5714721e882a893b051a49531bee6c8a847398