added a test of map 'set' operations.
Reminder: All three of the methods of SMRMap ,

	SMRMap::keyset
	SMRMap::valueSet
	SMRMap::entitySet

currently return -immutable copies- of the underlying map. As in Java, taking
the key/value/entry-set over a map which is concurrently being manipulated yield
undefined results. Any manipulation to the map –after- the set was snapshot does
not affect it. And operations like add/remove are not supported (it is
immutable).