Resolve and Persist Overloaded SMR Methods
In order to support method overloading for SMRObjects we need
to persist the method signature in the SMREntries instead
of the name only. When replaying SMREntries we need the parameter
types and return type to map to the correct method.

Also, since SMR objects allow for annotation naming of methods,
overloading conflicts are not detected during compile time, so
this checks for this conflicts during annotation processing.