Ensure that on nested transaction, optimistic updates from parent tx are properly applied.
Upon context switch (and creation of nested transaction), we have to sync the underlying object
to its optimistic state. To achieve that, we replay optimistic updates for each nested tx from
the root to the child. This fix also reverts changes to mergeWriteSetInto and remove write set
inheritance in transactions.