Remove Netty Memory Pools
Using ByteBuf(heap and direct) is causing memory leaks as a side
effect of pipelining problems. Until we re-architect some parts of
the pipeline, using pooled memory will affect system stability by
leaking memory. Since the performance enhancements of pooled memory
are not clear/necessary at this point, this patch removes it for
the sake of stability.