Bringing back to life the IRetry, turning the implementation upside (#528)
down

Preserved the approach with the generics used to transfer exceptions to the run method -
it is really cool (hack). A the exception management and auto decision whether we should
retry or not is cut with the big scissors:

1) The runnable now *explicitly* responsible to ask for retry by throwing RetryNeededException
2) All other exceptions are left to the user to handle them in the
standard, java way. The user is able to decide whether they need to
be handled from the lambda or outside, when run() throws them.
3) We don't swallow the InterruptedException any longer neither other exception
4) I added new features to ExponentialBackoffRetry, the base can be configured, and some portion of the
wait can be randomized, additionally a fixed wait can be appended to the timeout.
5) Added some unit tests