Add runtime's dependencies to Corfu-Repos Maven repo
Fixes #567.

Changes to SMR object annotations compilation also created
dependencies on Corfu JAR files outside of runtime's.
I'm not 100% sure that this small patch is sufficient; a Maven
repo wizard (@no2chem?) should review.

Prior to this patch, when using a `pom.xml` file as described
in the README (such as
https://gist.github.com/slfritchie/c02493a9aad7e69b143208148ba2b959)
results in the following Maven messages:

    [WARNING] The POM for org.corfudb:annotations:jar:0.1-2762 is missing, no dependency information available
    [WARNING] The POM for org.corfudb:annotationProcessor:jar:0.1-2762 is missing, no dependency information available

... and then ending with:

    [ERROR] Failed to execute goal on project test: Could not resolve dependencies for project org.corfudb.gh567:test:jar:0.01: The following artifacts could not be resolved: org.corfudb:annotations:jar:0.1-2762, org.corfudb:annotationProcessor:jar:0.1-2762: Could not find artifact org.corfudb:annotations:jar:0.1-2762 in corfu-mvn-repo (https://raw.github.com/CorfuDB/Corfu-Repos/mvn-repo/) -> [Help 1]