Corfu global log scanner & replica repair.
For all intervals in the global Corfu log address space,
the interval is either in a healthy state (as scanned by some
repair worker at an earlier time), or its state is unknown.
A SMRMap called 'globalState' is used to track healthy/unknown
state, in addition to a map of intervals where a worker process
is believed to be working on scanning & repairing data in that
interval.

A worker processes has an interval size that it prefers to
work with. (No need to use same size interval size with other
workers.)  A worker consults the unknown map and adds an
interval to the working map.  When the quorum repair scan
is complete, a worker removes the interval from working
and adds it to healthy.

Intervals that sit in the working map for too long are
considered failed and re-eligible to be scanned.  A reliable
detector of worker faults isn't needed: duplicate scans can't
damage Corfu data.