Bind to 127.0.0.1
When starting a server in single mode without specifying -a option
and a layout file, the server will automatically create a layout
for the localhost. As a consequence, the server will reply with
localhost addresses, this wont work for external clients. This
patch binds to 127.0.0.1 so that external connections cannot
connect to the server and thus it wont be able to reply with
localhost addresses.