Safe routers epoch update
Upon invalidateLayout, we should only set the client routers epoch to the
new layout epoch if they are greater than the current one they have. There
is no reason to get the routers to a smaller epoch. This prevents a downgrading
of epoch between seal and installation of a new Layout.