sequencer ready tests
cluster test currently disabled.

test currentlyfailing..

there seems to be a subtle race:

If a client grabs a token at epoch 1, then a server restarts at epoch 2, and the
client comes back to do addressSpaceView.write() using the old token, it will
loop forever:
                - write is rejected
                - layout will be invalidated (but it is already uptodate)
                - retry
                …

This commit introduces a fix to a stale-token bug and a test that exposes the
bug and verfies the fix.