A pair of handlers is provided in case of system unavailable. (#982)
NetworkExcpetion within the layoutHelper function will call the handler. The default
handler will do nothing and just let layoutHelper retry the request.

For writes(Transaction and single update), NetworkExceptions are passed to the
view layer (StreamsView and BackPointerView). After a NetworkException, the layer
will force a read to see if we persisted the record or not. If it was persisted,
the request return the correct value. If not, the retry logic is implemented in the view.