Object Open Race
When an object is opened, it is added to the cache right after
its creation, as a consequence other threads can retrieve the
object from the object cache after its creation and before
additional initialization. For example, register index. This
patch moves the extra initialization into an atomic operation
(i.e. computeIfAbsent). As a consequence, the parameters of
the first cached open will be observed by subsequent opens.