Use systemd 'control-group' as KillMode
Currently we are using 'process' as KillMode in service files.
If KillMode is set to process, only the main process itself is killed.
It will not kill child processes.

As per systemd documentation:

If KillMode is set to control-group, all remaining processes
in the control group of service will be killed on unit stop.
(First SIGTERM will be sent to all remaining service processes,
 then SIGKILL will be sent to all remaining service processes
 once TimeoutStopSec expired.)
 
https://www.freedesktop.org/software/systemd/man/systemd.kill.html

Issue: #2021025
Change-Id: Ia3c09dffdd7bf8b1a73eae0da802918859c79e25