Remove Node (#1151)
* Retry Actions On Failure

Unlike the client, the orchestrator service knows which particular
action failed, this patch enables the orchestrator to retry failed
actions thus decreasing the probability of a clients retrying and
saving the round trip, in effect increasing robustness.

* Remove Node

An implementation of a remove node. This workflow allows the removal
of a node from the cluster.