ransaction Abort Exception Cause for slow writers
The cause CONFLICT is reported whenever two transactions conflict over
the same keys for any given stream(s). This change corrects the overload
of this abort cause for the case of slow writers, i.e., whenever a
writer fails to write given that the designated address is continuously
overwritten (perhaps hole filled) by a faster reader. This is due to the
fact that it is not really conflicting with any other stream, the cause
of abort is that you are overwritten by a fast reader in the system.

We also consider a corner case which protects from inconsistency,
i.e., when the data is already present (case of a partial write
propagated by a fast reader).