Force Timeout On Connection (#1414)
The tcp connection can get in a state where the client doesn't
receive a termination signal, but the server dies abruptly. In this
case, the client can hang until the OS times out the connection.
This can take a very long time (in our case 15min) until the system
becomes available again.