Transactions Simulator
A parametrized model to simulate a transactional load on the cluster.
The simulator is immune to prematurely trimmed streams, so it doesn't
run a checkpoint/trimmer and doesn't care if its streams have been
trimmed. The parameters include, txn size, number of touched tables
in a transaction, the size of conflict keys etc.