Lookup by Index when Index is not specified
The use of the getByIndex API should fail whenever the index has not
been specified for the given table. This should be followed by the
proper exception and logging. In addition we add a message field to
the transaction aborted exception info in case a message for the cause
of abort is available.