Set corfu-server "niceness" value to -10
Description:
The idea is to priortize the critical services in NSX over others.
As a first step towards this goal, we have come up with two type of
service tiers. Tier 1 consists of the critical services in NSX
i.e. CCP, CBM and Corfu (note that CCP has a dependency on CBM and
Corfu and, therefore, they also need to be priorized with CCP). The
other tier consists of all the other services.

We are setting the "niceness" of these tier 1 services to -10 (priority
10). The defailt "niceness" is 0. Note that the critical kernel
services have a "niceness" of "-20" and, therefore, they should not be
impacted because of this change.

Refer to https://gitreview.eng.vmware.com/#/c/nsx/+/180975/ for the
corresponding change in the nsx repo.

Version: 0.2.2.20181207235326.2332

Change-Id: I33a74dd124447c409e02166e2bf8682df7d06e66