Corfu tanuki wrapper should not shutdown if the JVM doesn't terminate
Description:
Currently, if the JVM doesn't terminate within a timeout,
the wrapper aborts all restarts and shuts itself down. The wrapper
which is a watchdog should not shutdown like this. There are scenarios
where the JVM can't terminate within a timeout because of valid reasons
such a bad system thrash. We are changing this behavior so that
the wrapper never times out in such scenarios.

Issue: #2259717
Change-Id: I483fb9f3530d586c793fe11ab97da6ae0633bd53