MultiVersion Checkpoints
Instead of using the same snapshot to checkpoint all the maps
to be checkpointed, this patch makes the CheckpointWriter
use increasing checkpoint snapshots for every map and
therefore decreasing the probability of rolling back the stream.
Because multiple increasing snapshots are used, only the smallest
snapshot is used for trimming.