FillReadQueue Optimization
Fill from resolved queue when part of the data is already available, and
follow backpointers strictly on the addresses that are not part of the stream.

Remove checkpointSuccessEndAddr which is never used.

Persist checkpointSnapshotAddress to Log Metadata as this is used in the runtime
for determining the stop address on follow backpointers and deciding upon
a transaction being retriable or not.