Allow transactions that span epochs to commit if primary sequencer not changed. (#1771)
- Currently we abort any transaction that spans epochs. But this is not
necessary. If the primary sequencer remains the same (i.e. it has
consecutive epochs), we could allow the transaction with timestamp in
this epoch range to commit as the sequencer has all the information.
This could improve performance in case of flakey nodes.

- Removed sequencer thread configurations and all synchronization
modifiers in SequencerServer since it is single thread by design.