Add Priority Level For Clients
Added the ability for clients to tag their requests with a priority
level, this priority is then used by the servers to reject certain
tasks under resource constraints.

Added the ability to set a max size on the data log. The LogUnit
will now track the log size and fail writes after the log size
has been reached. Except for the case where the writes originate
from a high priority client.

Ignore QuorumReplicationProtocolTest Tests

Set the priority level of the management clients to high.

Log the CorfuRuntimeParameters when the client connects.

readingTrimmedAddress wait for futures.

refactored initStreamLogDirectory