Do not use stale conflictKey/conflictStream on overwritten for TXNs.
On OverwriteException, StreamsView::append will grab a new token and
retry, but it was still using the conflictKey/conflictStream info from
the stale tokeResponse. In case of transactions and there is a conflict
when we try to re-requst the token, the stale conflictKey/conflictStream
is empty. This patch directly uses the new token and retry.