MultiCheckpointWriter Bug Fix
Whenever the last address of a stream is a hole (right before the checkpointer is triggered)
We build the checkpoint start log address based on the wrong snapshot of the stream,
this is because we discard holes and hence set the start to the previous DATA entry.
This will diverge from the state kept by the sequencer for this stream on trims (trim mark).