ILogData: Get rid of backward dependency on LogEntry.
LogEntry is the payload of LogData, however LogEntry also contains a
reference to LogData, which undermines readability and is error prone.
This is also unnecessary because the only reason to have this backward
dependency is LogEntry needs to know about the global address inside
LogData, which is consumed by VLO. This patch removed this dependency
by setting globalAddress directly.