Use A Reverse Iterator
As more entries are added to a stream, the stream tail query (i.e
getAddressesInRange) operation becomes slower and slower, because
the whole bit set has to be iterated. This happens despite the fact
that clients have most of the StreamSpace resolved. This patch
improves the performance by iterating backwards and not iterate
more addresses than needed.