GC: Unify LogEntry types: SMREntry, MultiSMREntry, MultiObjectSMREntry.
We have three SMR LogEntry types, namely SMREntry for a non-transactional
update, MultiObjectSMREntry for transactional updates and MultiSMREntry
for checkpoint entry. This creates a lot of confusion, boilerplates and
difficulties when handling log updates. With this patch, all transactional,
non-transactional updates and checkpoint entries are composed of a single
log entry type: SMRLogEntry, which is similar to MultiObjectSMREntry. And
each SMR update is turned into SMRRecord, which is similar to SMREntry,
but is not inherited from LogEntry any more, because now the only LogEntry
type persisted on log unit server would be SMRLogEntry.