Use global compaction mark instead of per stream to prevent leak.
Using per steam compaction mark has a downside of storing compacted
stream ids in each LogData's metadata, because each read on an address
needs to return the per stream compaction marks on that address, in
this way, we cannot completely remove a LogData which causes a leak.
Using global compaction mark, we only needs to store a bitmap for the
compacted addresses in a segment, which is a smaller overhead, but
sacrifices some versioning ability. This is fine since we have a policy
to ensure compaction mark does not move for a certain amount of time.
This patch also adds compaction for holes.