CorfuGuidGenerator: Prevent transactional aborts
If CorfuQueue is used in a highly concurrent environment, it
causes the correction to quickly overflow and bump up the instance id.
Bumping up the instance id is really a corfu transaction.
If this transaction is nested, it aborts the parent transaction.
To avoid this, keep the use of transactions to an absolute minimum.
Make CorfuGuids tolerant of NTP time drifts that are under a minute.

Issue: #2152