GC: Introducing auto commit mechanism. (#2176)
* GC: Introducing auto commit mechanism.

- Added an auto commit service in management server, which periodically
checks if the address prefix is committed, if not then hole fill these
addresses. This makes the entire log prefix consolidated.

- With auto commit, compactor only compact on the consolidated prefix,
so that we don't need to persist the compacted addresses for each segment
in order to differentiate a hole and a compacted address. This improves
disk space usage in longevity.

- Also with auto commit, compactor don't need to worry about syncing the
delta updates on a segment if it's currently being compacted since the
segment is already consolidated.

* Use requreStateTransfer flag on log unit server.