GC: Update multi write sanity check on log unit.
The range write API needs to check the range is contiguous and does not
span two segments. With sparse trim GC, this will not be the case since
the addresses to write are sparse, we just need to check if the entries
to append are ordered by global address and contain no duplicates.