CorfuStore Streaming shouldn't leak streamViews
CorfuStore just has 1 Transaction poller. But to ensure that new
subscribers are taken care of, periodically this transaction poller
is refreshed and a new one is created. But this causes a leak in the
underlying runtime's TRANSACTION StreamsView's openedStreams
resuling in the JVM going out of Memory after a while.
Simple fix is to get an unsafe reference so JVM GC takes care of leak.