Remove Parent Filesystem read-only check for LogUnitServer
Starting with Mac OS X 10.15 the parent filesystem returned
by the System.getProperty("java.io.tmpdir") is marked as
read-only. However this does not mean that the returned
temporary directory itself is read-only.
So given that this can happen in production as well,
removing the redundant parent filesystem read-only
check would prevent false panics and allow IT tests
to run on Mac as well.