CorfuStore: Record if table is opened as disk based
This is needed for checkpointer so it can do the same.

CorfuBrowser: Add support for disk tables
Without it browser can't open large tables.
If a table is too large to fit in memory then user can
pass a --diskPath and let that table be
opened as an in-memory table.

CorfuStore: Add test to checkpoint & Trim disk based
CorfuStore table Without this bugs are only
discovered when deployed in scale setups.