CorfuQueue: Sort entries before returning in entryList()
This is to preserve order if checkpointer runs first.
Checkpointer uses a normal HashMap and not a LinkedHashMap
causing a loss of ordering if it runs first.
Fix is to sort the entries first.
This also enables the api to allow entries to be returned
only after a user specified previous entry.