Fix LogData lastKnownSize which is inaccurate and inconsistent.
- When set during serialization and deserialization, the LogData
lastKnownSize is set to different values. It is also inaccurate
as it is set to ByteBuf's internal array length during serialization,
which is usually larger the actually (it is dynamically resized and
usually 2x of previous size). It needs to be set to the payload's
serialized size. This incorrect behavior affects transaction size
limit calculation, which results in unnecessary WriteSizeException.

- Also fixed a test where compression should be ruled out.