Fix Trim Exception on Raw Stream Access
When a stream is being consumed directly (remaining/remainingUpTo
in the StreamViews), the readCpQueue is not being cleared.

If on the first access there is a valid CP to load from, the addresses
belonging to it will always remain in memory in the readCPQueue.
If other entries are found, these will be read along the ones in the readCPQueue (always,
as they're not cleared), when they're trimmed we will throw a TrimmedException on an old CP.